requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
aioconsole==0.1.15        # via aiomonitor
aiomonitor==0.4.5         # via -r requirements-dev.in
appdirs==1.4.3            # via black, virtualenv
appnope==0.1.0            # via ipython
aspy.yaml==1.3.0          # via pre-commit
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black
autoflake==1.3.1          # via -r requirements-dev.in
autopep8==1.5             # via -r requirements-dev.in, pathlib-mate
backcall==0.1.0           # via ipython
betamax-matchers==0.4.0   # via -r requirements-dev.in
betamax==0.8.1            # via -r requirements-dev.in, betamax-matchers
black==19.10b0            # via -r requirements-dev.in
certifi==2020.4.5.1       # via requests
cfgv==3.0.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.0                # via black
colorama==0.4.3           # via hunter
decorator==4.4.1          # via ipython, traitlets
distlib==0.3.0            # via virtualenv
docutils==0.16            # via restructuredtext-lint
fancycompleter==0.9.1     # via pdbpp
filelock==3.0.12          # via virtualenv
freezegun==0.3.14         # via -r requirements-dev.in
hunter==3.1.3             # via -r requirements-dev.in
identify==1.4.11          # via pre-commit
idna==2.9                 # via requests
importlib-metadata==1.5.0  # via pre-commit, virtualenv
invoke==1.4.1             # via -r requirements-dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.12.0           # via -r requirements-dev.in, pystuck
isort==4.3.21             # via -r requirements-dev.in, pylint
jedi==0.16.0              # via ipython
lazy-object-proxy==1.4.3  # via -r requirements-dev.in, astroid
libcst==0.3.5             # via monkeytype
manhole==1.6.0            # via hunter
mccabe==0.6.1             # via -r requirements-dev.in, pylint
monkeytype==20.5.0        # via -r requirements-dev.in
mypy-extensions==0.4.3    # via monkeytype, mypy, typing-inspect
mypy==0.770               # via sqlalchemy-stubs
nodeenv==1.3.5            # via pre-commit
notify-send==0.0.13       # via -r requirements-dev.in
parso==0.6.1              # via jedi
pathlib-mate==0.0.15      # via -r requirements-dev.in
pathspec==0.7.0           # via black
pathtools==0.1.2          # via watchdog
pdbpp==0.10.2             # via -r requirements-dev.in
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
plumbum==1.6.8            # via rpyc
pre-commit==2.0.1         # via -r requirements-dev.in
prompt-toolkit==3.0.3     # via ipython
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.5.0        # via autopep8
pyflakes==2.1.1           # via autoflake
pygments==2.5.2           # via ipython, pdbpp
pylint==2.4.4             # via -r requirements-dev.in
pyrepl==0.9.0             # via fancycompleter
pystuck==0.8.5            # via -r requirements-dev.in
python-dateutil==2.8.1    # via freezegun
pyyaml==5.3               # via aspy.yaml, libcst, pre-commit
regex==2020.1.8           # via black
requests-toolbelt==0.9.1  # via betamax-matchers
requests==2.23.0          # via betamax, requests-toolbelt
restructuredtext-lint==1.3.0  # via -r requirements-dev.in
rope==0.17.0              # via -r requirements-dev.in
rpyc==4.1.4               # via pystuck
six==1.14.0               # via astroid, freezegun, pathlib-mate, python-dateutil, traitlets, virtualenv
sqlalchemy-stubs==0.3     # via -r requirements-dev.in
terminaltables==3.1.0     # via aiomonitor
toml==0.10.0              # via black, pre-commit
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via astroid, black, mypy
typing-extensions==3.7.4.2  # via libcst, mypy, sqlalchemy-stubs, typing-inspect
typing-inspect==0.6.0     # via libcst
unittest2pytest==0.4      # via -r requirements-dev.in
urllib3==1.25.9           # via requests
virtualenv==20.0.2        # via pre-commit
vulture==1.3              # via -r requirements-dev.in
watchdog==0.10.2          # via when-changed
wcwidth==0.1.8            # via prompt-toolkit
when-changed==0.3.0       # via -r requirements-dev.in
wmctrl==0.3               # via pdbpp
wrapt==1.11.2             # via astroid
zipp==2.2.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
